+++ windows/vs10/povclr/RenderTarget.h
@@ -5,7 +5,7 @@ namespace povclr
 	public interface class RenderTarget
 	{
 		void Initialize(int width, int height);
-		void SetPixel(int x, int y, double red, double green, double blue, double alpha);
+		void SetPixel(int x, int y, float red, float green, float blue, float alpha);
 		void Finish();
 	};
 }
